+++ src/pipeline.h
@@ -187,11 +187,12 @@ struct PipelineBaton {
   bool jxlLossless;
   VipsBandFormat rawDepth;
   std::string err;
-  bool withMetadata;
+  int keepMetadata;
   int withMetadataOrientation;
   double withMetadataDensity;
-  std::string withMetadataIcc;
-  std::unordered_map<std::string, std::string> withMetadataStrs;
+  std::string withIccProfile;
+  std::unordered_map<std::string, std::string> withExif;
+  bool withExifMerge;
   int timeoutSeconds;
   std::unique_ptr<double[]> convKernel;
   int convKernelWidth;
@@ -353,9 +354,10 @@ struct PipelineBaton {
     jxlEffort(7),
     jxlLossless(false),
     rawDepth(VIPS_FORMAT_UCHAR),
-    withMetadata(false),
+    keepMetadata(0),
     withMetadataOrientation(-1),
     withMetadataDensity(0.0),
+    withExifMerge(true),
     timeoutSeconds(0),
     convKernelWidth(0),
     convKernelHeight(0),